Add Movie class to model for Day 3 challenge
The Movie class is a Plain old Java Object (POJO) with no setters and
only getters. This modeling was chosen as to represent the movie entity
as simple as possible. As we are going to use Jackson to unmarshall the
JSON returned from the API call, the getters are necessary, but, as the
entity should be imutable, no setters are implemented.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>