Add implementation for Day 5 challenge
This commit splits the HTTP call to the IMDB API and the parsing of the
JSON returned into two classes: ImdbApiClient and ImdbMoviesJsonParser.
As a consequence, this commit also adapts the entrypoint class (App
class) and changes the simple test suite from testing the now gone
ImdbApiConsumer class, to the ImdbMoviesJsonParser class (as the only
behaviour that was tested revolved around the parsing of the JSON).

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>